added login and register layout
