+++ app/layout.tsx
@@ -6,8 +6,8 @@ import AuthProvider from "./context/AuthProvider";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "TODO",
+  description: "TODO APP",
 };
 
 export default function RootLayout({
